backend/requirements.txt
fastapi
uvicorn
fredapi
python-dotenv
requests
